chore: move db files into new storage module
