+++ twenty-first/src/lib.rs
@@ -1,6 +1,7 @@
 #![deny(clippy::shadow_unrelated)]
 pub mod amount;
 pub mod shared_math;
+pub mod storage;
 pub mod test_shared;
 pub mod timing_reporter;
 pub mod util_types;
