+++ twenty-first/src/util_types.rs
@@ -1,9 +1,6 @@
 pub mod algebraic_hasher;
 pub mod blake3_wrapper;
-pub mod database_array;
-pub mod database_vector;
 pub mod emojihash_trait;
-pub mod level_db;
 pub mod merkle_tree;
 pub mod merkle_tree_maker;
 pub mod mmr;
